+++ tests/test_query.py
@@ -8,7 +8,7 @@ import pytest
 from astropy.coordinates import SkyCoord, Angle
 from mocpy import MOC
 from pytest_mock import mocker, MockerFixture  # noqa: F401
-from typing import List
+from typing import List, Union
 
 import vasttools.query as vtq
 
@@ -684,6 +684,41 @@ class TestQuery:
             ' are found in the VAST Pilot survey footprint!'
         )
 
+    def test_init_failure_invalid_scheduler(self,
+                                           mocker: MockerFixture
+                                           ) -> None:
+        """
+        Tests the initialisation failure of a Query object.
+
+        Specifically when the requested dask scheduler is invalid.
+
+        Args:
+            mocker: The pytest-mock mocker object.
+
+        Returns:
+            None
+        """
+        isdir_mocker = mocker.patch(
+            'vasttools.query.os.path.isdir',
+            return_value=True
+        )
+        mocker_data_available = mocker.patch(
+            'vasttools.query.Query._check_data_availability',
+            return_value=True
+        )
+
+        with pytest.raises(vtq.QueryInitError) as excinfo:
+            query = vtq.Query(
+                planets=['Mars'],
+                scheduler='bad-option',
+                base_folder='/testing/folder'
+            )
+
+        assert str(excinfo.value) == (
+            "bad-option is not a suitable scheduler option. Please "
+            "select from ['processes', 'single-threaded']"
+        )
+
     def test_init_settings(self, mocker: MockerFixture) -> None:
         """
         Tests the initialisation of a Query object.
@@ -721,6 +756,7 @@ class TestQuery:
         forced_cluster_threshold = 7.5
         output_dir = '/output/here'
         incl_observed = False
+        scheduler = 'processes'
 
         expected_settings = {
             'epochs': ["1", "2", "3x"],
@@ -737,7 +773,8 @@ class TestQuery:
             'output_dir': output_dir,
             'search_around': False,
             'tiles': use_tiles,
-            'incl_observed': False
+            'incl_observed': False,
+            'scheduler': 'processes'
         }
 
         query = vtq.Query(
@@ -756,7 +793,8 @@ class TestQuery:
             forced_allow_nan=forced_allow_nan,
             forced_cluster_threshold=forced_cluster_threshold,
             output_dir=output_dir,
-            incl_observed=incl_observed
+            incl_observed=incl_observed,
+            scheduler=scheduler
         )
 
         assert query.settings == expected_settings
@@ -1618,3 +1656,64 @@ class TestQuery:
         assert vast_query_psrj2129.vast_p1 == vast_p1
         assert vast_query_psrj2129.vast_p2 == vast_p2
         assert vast_query_psrj2129.vast_full == vast_full
+
+    @pytest.mark.parametrize("req_epochs, epochs_expected",
+                             [("0", ["0"]),
+                              ("3x", ["3x"]),
+                              ("3", ["3x"]),
+                              ("0,1", ["0", "1"]),
+                              ("0,3x", ["0", "3x"]),
+                              ("0,3", ["0", "3x"]),
+                              ([0,1], ["0", "1"]),
+                              (0, ["0"]),
+                              (["0", "1"], ["0", "1"]),
+                              (["0", 1], ["0", "1"]),
+                              ([1, "3x"], ["1", "3x"]),
+                              ([1, "3"], ["1", "3x"]),
+                              ("all", ["0", "1", "3x"]),
+                              ("all-vast", ["1", "3x"]),
+                              ],
+                             ids=('single-str',
+                                  'single-str-x-provided',
+                                  'single-str-x-missing',
+                                  'multiple-str',
+                                  'multiple-str-x-provided',
+                                  'multiple-str-x-missing',
+                                  'single-int',
+                                  'int-list',
+                                  'str-list',
+                                  'mixed-list',
+                                  'mixed-list-x-provided',
+                                  'mixed-list-x-missing',
+                                  'all',
+                                  'all-vast'
+                                  )
+                              )
+    
+    def test__get_epochs(
+        self,
+        vast_query_psrj2129: vtq.Query,
+        req_epochs: Union[str, List[str], List[int]],
+        epochs_expected,
+        mocker: MockerFixture,
+        ) -> None:
+        """
+        Test the get_epochs function.
+        
+        Args:
+            vast_query_psrj2129: The dummy Query instance that includes
+                a search for PSR J2129-04 with the included found fields data.
+            req_epochs: The requested epochs.
+            epochs_expected: The expected output of the function.
+            
+        Returns:
+            None
+        """
+        mocked_released_epochs = {"0":"00", "1":"01", "3x":"03x"}
+        
+        mocker.patch("vasttools.query.RELEASED_EPOCHS",
+                     new=mocked_released_epochs
+                     )
+        
+        returned_epochs = vast_query_psrj2129._get_epochs(req_epochs)
+        assert returned_epochs == epochs_expected