+++ vasttools/bin/find_sources.py
@@ -199,6 +199,13 @@ def parse_args() -> argparse.Namespace:
         action="store_true",
         help=("Overwrite the output directory if it already exists."))
     parser.add_argument(
+        '--scheduler',
+        default='processes',
+        choices=['processes', 'single-threaded'],
+        help=("Dask scheduling option to use. Options are 'processes' "
+              "(parallel processing) or 'single-threaded'.")
+        )
+    parser.add_argument(
         '--sort-output',
         action="store_true",
         help=(
@@ -508,7 +515,8 @@ def main() -> None:
         forced_cluster_threshold=args.forced_cluster_threshold,
         forced_allow_nan=args.forced_allow_nan,
         incl_observed=args.find_fields,
-        corrected_data=not args.uncorrected_data
+        corrected_data=not args.uncorrected_data,
+        scheduler=args.scheduler,
     )
 
     if args.find_fields: